fix(*): implemented lazy loading on auth screens
